Omni AI/requirements.txt
PyQt6
PyAutoGUI
pyyaml
requests
aiohttp
apscheduler
pyttsx3
speechrecognition
google-generativeai
python-dotenv
sqlite3
psutil
